+++ extensions/visual-studio-code-recent-projects/src/open-new-window.ts
@@ -2,6 +2,8 @@ import { Toast, closeMainWindow, showToast } from "@raycast/api";
 import { runAppleScript } from "@raycast/utils";
 import { build } from "./preferences";
 import { VSCodeBuild } from "./types";
+import { isMacOs, isWin } from "./utils";
+import { getVSCodeCLI } from "./lib/vscode";
 
 /**
  * The index of the `New Window` menu item in the `File` menu.
@@ -32,7 +34,7 @@ const NewWindowMenuItemIndex: Record<VSCodeBuild, number> = {
  * However, for Cursor, which does not have a `New File` menu item, `New Window` is in the second position.
  * We need to handle this case specially.
  */
-const makeNewWindow = async () => {
+const makeNewWindowMacos = async () => {
   await runAppleScript(`
     tell application "${build}"
 	    activate
@@ -56,10 +58,22 @@ const makeNewWindow = async () => {
   `);
 };
 
+const makeNewWindowWindows = async () => {
+  const cli = getVSCodeCLI();
+  cli.newWindow();
+};
+
 export default async function command() {
   try {
     await closeMainWindow();
-    await makeNewWindow();
+
+    if (isWin) {
+      await makeNewWindowWindows();
+    }
+
+    if (isMacOs) {
+      await makeNewWindowMacos();
+    }
   } catch (error) {
     await showToast({
       title: "Failed opening new window",